Add features to list of phones

This function counts occurrences of phones and includes basic phonetic features.

Usage

featurise(phlist)

Arguments

phlist

A list of phones or the output of phonetise().

Value

A tibble.

Examples

ipa <- c("ada", "buba", "kiki", "sa\u0283a")
ip_ph <- phonetise(ipa)
featurise(ip_ph)
#> # A tibble: 8 × 17
#>   phone count base  unicode uni_name  ipa_name phon_type type  height_ipa height
#>   <chr> <int> <chr> <chr>   <chr>     <chr>    <chr>     <chr> <chr>      <chr> 
#> 1 d         1 d     U+0064  latin sm… voiced … base      cons… NA         NA    
#> 2 s         1 s     U+0073  latin sm… voicele… base      cons… NA         NA    
#> 3 u         1 u     U+0075  latin sm… close b… base      vowel close      high  
#> 4 ʃ         1 ʃ     U+0283  latin sm… voicele… base      cons… NA         NA    
#> 5 b         2 b     U+0062  latin sm… voiced … base      cons… NA         NA    
#> 6 i         2 i     U+0069  latin sm… close f… base      vowel close      high  
#> 7 k         2 k     U+006B  latin sm… voicele… base      cons… NA         NA    
#> 8 a         5 a     U+0061  latin sm… open fr… base      vowel open       low   
#> # ℹ 7 more variables: backness <chr>, rounding <chr>, voicing <chr>,
#> #   place <chr>, manner <chr>, lateral <lgl>, sonorant <lgl>